requirements.txt
pandas
matplotlib
seaborn
scikit-learn
xgboost
statsmodels
tensorflow
PyYAML
holidays
streamlit